test: move get-all-votes inside results page
