+++ components/poll-container.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import TheNewStackLogo from './the-new-stack-logo';
 
 export default function PollContainer({ children }: { children: React.ReactNode }) {
